CORE-5114 Move WireTransactionSerializationTest to ledger-common-impl and make some kryoserialization accessible for it
