+++ libs/ledger/ledger-common-impl/cpks/empty/build.gradle
@@ -1,0 +1,27 @@
+plugins {
+    id 'net.corda.plugins.cordapp-cpb'
+}
+
+description 'Corda Common Ledger impl empty CPK'
+
+group 'com.example.sandbox'
+version ''
+
+ext {
+    platformVersion = 999
+}
+
+cordapp {
+    targetPlatformVersion platformVersion as Integer
+    minimumPlatformVersion platformVersion as Integer
+    workflow {
+        name "Corda Common Ledger impl empty CPK"
+        versionId 1
+        vendor "R3"
+    }
+}
+
+dependencies {
+    cordaProvided platform("net.corda:corda-api:$cordaApiVersion")
+    cordaProvided 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
+}
