+++ libs/serialization/serialization-kryo/src/main/java/net/corda/kryoserialization/package-info.java
@@ -1,4 +1,6 @@
 @QuasarIgnoreAllPackages
+@Export
 package net.corda.kryoserialization;
 
 import co.paralleluniverse.quasar.annotations.QuasarIgnoreAllPackages;
+import org.osgi.annotation.bundle.Export;
