+++ libs/serialization/serialization-kryo/src/main/java/net/corda/kryoserialization/resolver/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.kryoserialization.resolver;
+
+import org.osgi.annotation.bundle.Export;
