+++ libs/serialization/serialization-kryo/src/main/java/net/corda/kryoserialization/serializers/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.kryoserialization.serializers;
+
+import org.osgi.annotation.bundle.Export;
