+++ settings.gradle
@@ -229,6 +229,7 @@ include 'libs:layered-property-map'
 include 'libs:layered-property-map:layered-property-map-test-converter'
 include 'libs:ledger:notary'
 include 'libs:ledger:ledger-common-impl'
+include 'libs:ledger:ledger-common-impl:cpks:empty'
 include 'libs:ledger:ledger-common-serialization'
 include 'libs:ledger:ledger-consensual-impl'
 include 'libs:lifecycle:lifecycle'
